Move fixture to src folder
