+++ src/index.ts
@@ -1,6 +1,6 @@
 import pwApi from "./pwApi";
 import axiosApi from "./axiosApi";
 
-const { testApi: test } = require('./fixtures/testApi.js');
+const { testApi: test } = require('./testApi.js');
 
 export { pwApi, axiosApi, test }
